+++ apps/blog/src/app/admin/page.tsx
@@ -1,11 +1,11 @@
 'use client';
 
 import { Button } from '@joseph0926/ui/components/button';
+import { Input } from '@joseph0926/ui/components/input';
 import { Plus, Search } from 'lucide-react';
 import { useState } from 'react';
 import { PostDialog } from '@/components/admin/post-dialog';
 import { PostsTable } from '@/components/admin/posts-table';
-import { Input } from '@/components/ui/input';
 
 export default function AdminPage() {
   const [searchQuery, setSearchQuery] = useState('');
