+++ apps/blog/src/components/loading/post-header.loading.tsx
@@ -1,4 +1,4 @@
-import { Skeleton } from '../ui/skeleton';
+import { Skeleton } from '@joseph0926/ui/components/skeleton';
 
 export const PostHeaderLoading = () => {
   return <Skeleton className="h-80 w-full" />;
