+++ packages/ui/components/form.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import { Label } from '@joseph0926/ui/components/label';
 import { cn } from '@joseph0926/ui/lib/utils';
 import * as LabelPrimitive from '@radix-ui/react-label';
 import { Slot } from '@radix-ui/react-slot';
@@ -13,7 +14,6 @@ import {
   useFormContext,
   useFormState,
 } from 'react-hook-form';
-import { Label } from '@/components/ui/label';
 
 const Form = FormProvider;
