+++ packages/ui/package.json
@@ -32,9 +32,15 @@
   },
   "packageManager": "pnpm@10.15.1",
   "dependencies": {
+    "@hookform/resolvers": "^4.1.3",
     "@radix-ui/react-dialog": "^1.1.15",
     "@radix-ui/react-dropdown-menu": "^2.1.16",
+    "@radix-ui/react-label": "^2.1.7",
+    "@radix-ui/react-popover": "^1.1.15",
     "@radix-ui/react-slot": "^1.2.3",
-    "cmdk": "^1.1.1"
+    "@radix-ui/react-tabs": "^1.1.13",
+    "cmdk": "^1.1.1",
+    "react-hook-form": "^7.62.0",
+    "zod": "^3.25.76"
   }
 }
